errors correction
