+++ db/migrate/20190618131340_add_category_id_to_notes.rb
@@ -1,0 +1,5 @@
+class AddCategoryIdToNotes < ActiveRecord::Migration[5.2]
+  def change
+    add_reference :notes, :category, foreign_key: true
+  end
+end
