+++ db/migrate/20190618131355_add_user_id_to_notes.rb
@@ -1,0 +1,5 @@
+class AddUserIdToNotes < ActiveRecord::Migration[5.2]
+  def change
+    add_reference :notes, :user, foreign_key: true
+  end
+end
