+++ db/migrate/20190618131408_add_favorite_to_notes.rb
@@ -1,0 +1,5 @@
+class AddFavoriteToNotes < ActiveRecord::Migration[5.2]
+  def change
+    add_column :notes, :favorite, :string
+  end
+end
